requirements.txt
aioice==0.9.0
aiortc==1.5.0
altair==5.0.1
attrs==23.1.0
av==10.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cmake==4.0.2
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
decorator==5.1.1
dnspython==2.4.1
easyocr==1.7.0
filelock==3.12.2
fonttools==4.41.1
fsspec==2024.6.1
gitdb==4.0.10
GitPython==3.1.32
google-crc32c==1.5.0
idna==3.4
ifaddr==0.2.0
imageio==2.31.1
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lazy_loader==0.3
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyclipper==1.3.0.post4
pycparser==2.21
pydeck==0.8.0
pyee==11.0.0
Pygments==2.15.1
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
scikit-image==0.21.0
scipy==1.11.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.25.0
streamlit-webrtc==0.45.1
sympy==1.12
tenacity==8.2.2
tifffile==2023.7.18
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2+cu118
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.143
urllib3==2.0.4
validators==0.20.0
watchdog==3.0.0
zipp==3.16.2
